requirements.txt
streamlit
pillow
inference-sdk
osmnx
networkx
plotly
geopy
hashlib
pandas
numpy
